transferred watch connectivity to InterfaceController
